% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nesterov_algoritm.R
\name{nesterov_gradient_descent}
\alias{nesterov_gradient_descent}
\title{nesterov_gradient_descent}
\usage{
nesterov_gradient_descent(f, alpha, x0, iterations, beta)
}
\arguments{
\item{beta}{}
}
\description{
Implémentation de la descente de gradient avec Nesterov
}
\examples{
gradient  <- nesterov_gradient_descent(f, alpha, x0, iterations, beta)
}
